tests/baseline/protocol-1.2/inspector/console/console-format-style-whitelist-expected.txt

CONSOLE MESSAGE: line 8: %cColors are awesome.
CONSOLE MESSAGE: line 9: %cSo are fonts!
CONSOLE MESSAGE: line 10: %cAnd borders and margins and paddings!
CONSOLE MESSAGE: line 11: %ctext-* is fine by us!
CONSOLE MESSAGE: line 13: %cDisplay, on the other hand, is bad news.
CONSOLE MESSAGE: line 14: %cAnd position too.
Tests that console logging dumps properly styled messages, and that the whole message gets the same style, regardless of multiple %c settings.

console-format-style-whitelist.html:4 Colors are awesome.
Styled text #0: color: blue;
console-format-style-whitelist.html:5 So are fonts!
Styled text #0: font-style: normal; font-variant: normal; font-weight: normal; font-stretch: normal; font-size: 1em; line-height: normal; font-family: Helvetica;
console-format-style-whitelist.html:6 And borders and margins and paddings!
Styled text #0: border: 1px solid red; margin: 20px; padding: 10px;
console-format-style-whitelist.html:7 text-* is fine by us!
Styled text #0: text-decoration-line: none; text-decoration-style: initial; text-decoration-color: initial;
console-format-style-whitelist.html:9 Display, on the other hand, is bad news.
Styled text #0: NO STYLES DEFINED
console-format-style-whitelist.html:10 And position too.
Styled text #0: NO STYLES DEFINED
